More code isolation. And cleanup
