+++ Plugins/SequencePlugin/Source/SequencePlugin/Private/Util/HexUtility.h
